# back/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HolterApiKeyOut": {
        "description": "La API key en claro de un equipo. Solo la ve un admin.\n\nLa devuelven tanto el alta y la rotación como `GET /devices/{id}/api-key`:\nla key se guarda cifrada además de hasheada, así que se puede volver a leer.\n`rotatedAt` es cuándo se generó la que se está devolviendo.",
        "properties": {
          "apiKey": {
            "title": "Apikey",
            "type": "string"
          },
          "deviceId": {
            "format": "uuid",
            "title": "Deviceid",
            "type": "string"
          },
          "rotatedAt": {
            "format": "date-time",
            "title": "Rotatedat",
            "type": "string"
          },
          "serial": {
            "title": "Serial",
            "type": "string"
          }
        },
        "required": [
          "deviceId",
          "serial",
          "apiKey",
          "rotatedAt"
        ],
        "title": "HolterApiKeyOut",
        "type": "object"
      },
...
      "PatientPasswordOut": {
        "description": "Contraseña nueva, también entregada una sola vez.",
        "properties": {
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "PatientPasswordOut",
        "type": "object"
      },
...
      "PatientStudyStatus": {
        "enum": [
          "active",
          "completed",
          "paused",
          "none"
        ],
        "title": "PatientStudyStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "Holter ECG API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/devices/{device_id}/api-key": {
      "get": {
        "description": "Devuelve la API key en claro del equipo, para grabarla en su firmware.\n\nSolo admin: la key habilita a subir señal como ese dispositivo, así que\nentregarla es equivalente a entregar el equipo. Cada lectura queda auditada.",
        "operationId": "get_holter_api_key_devices__device_id__api_key_get",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Device Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HolterApiKeyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Holter Api Key",
        "tags": [
          "devices"
        ]
      },
      "post": {
        "description": "Rota la API key del equipo: genera una nueva y deja la anterior inútil.\n\nSolo admin, por lo mismo que el GET. Es inmediato y no se puede deshacer —\nel chaleco que tenga cargada la key vieja empieza a recibir 401.",
        "operationId": "rotate_holter_api_key_devices__device_id__api_key_post",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Device Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HolterApiKeyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rotate Holter Api Key",
        "tags": [
          "devices"
        ]
      }
    },
...
    "/patients": {
      "get": {
        "operationId": "list_patients_patients_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "hasDevice",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hasdevice"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/PatientStudyStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "status[]",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/PatientStudyStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status[]"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 250,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "name",
              "pattern": "^(name|lastDataReceivedAt)$",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "asc",
              "pattern": "^(asc|desc)$",
              "title": "Order",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Patients",
        "tags": [
          "patients"
        ]
      },
      "post": {
        "description": "Alta del paciente **y** de su cuenta en la app móvil.\n\nLa respuesta trae `generatedPassword` una sola vez: Auth0 guarda el hash y\nno hay endpoint que la pueda volver a leer. El portal la muestra en un\ndiálogo con botón de copiar antes de que se pierda.",
        "operationId": "create_patient_patients_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientCreateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Patient",
        "tags": [
          "patients"
        ]
      }
    },
    "/patients/{patient_id}": {
      "delete": {
        "operationId": "delete_patient_patients__patient_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Patient",
        "tags": [
          "patients"
        ]
      },
      "get": {
        "operationId": "get_patient_patients__patient_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Patient",
        "tags": [
          "patients"
        ]
      },
      "patch": {
        "operationId": "update_patient_patients__patient_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Patient",
        "tags": [
          "patients"
        ]
      }
    },
    "/patients/{patient_id}/app-account": {
      "post": {
        "description": "Crea el acceso a la app de un paciente que todavía no lo tiene.\n\nPara las fichas cargadas antes de que la app existiera: la migración no les\ncreó cuenta retroactivamente porque muchas ni siquiera tienen email.",
        "operationId": "create_patient_app_account_patients__patient_id__app_account_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientPasswordOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Patient App Account",
        "tags": [
          "patients"
        ]
      }
    },
    "/patients/{patient_id}/app-password": {
      "post": {
        "description": "Genera una contraseña nueva y la devuelve en claro **una sola vez**.\n\nMismo criterio que `POST /devices/{id}/api-key`. No existe un endpoint para\n\"ver\" la contraseña actual porque no existe la contraseña actual: Auth0\nguarda el hash, y guardarla nosotros sería dejar credenciales de pacientes\nrecuperables desde la base.",
        "operationId": "regenerate_patient_app_password_patients__patient_id__app_password_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientPasswordOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate Patient App Password",
        "tags": [
          "patients"
        ]
      }
    },
    "/patients/{patient_id}/device": {
      "get": {
        "operationId": "get_patient_device_patients__patient_id__device_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HolterHealthOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Patient Device",
        "tags": [
          "patients"
        ]
      }
    },
    "/patients/{patient_id}/password-reset": {
      "post": {
        "description": "Manda el mail de recuperación de Auth0 al paciente.",
        "operationId": "send_patient_password_reset_patients__patient_id__password_reset_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Patient Password Reset",
        "tags": [
          "patients"
        ]
      }
    },
    "/patients/{patient_id}/studies": {
      "get": {
        "operationId": "get_patient_studies_patients__patient_id__studies_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientStudiesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Patient Studies",
        "tags": [
          "patients"
        ]
      }
    },
...
  }
}